finish the slider and grid
