+++ style/theme/default/compoments/avatar.less
