+++ style/theme/default/compoments/index.less
@@ -26,6 +26,7 @@
 @import './tooltip.less';
 @import './steps.less';
 @import './inputNumber.less';
+@import './slider.less';
