+++ style/theme/default/compoments/slider.less
@@ -1,23 +1,17 @@
-@import "../../colors/colors.less";
 
-@primary-color          : @green-6;
-@primary-3: color(~`colorPalette("@{primary-color}", 3)`);
-@primary-4: color(~`colorPalette("@{primary-color}", 4)`);
-@background-color-base  : hsv(0, 0, 96%);
-@border-color-split     : hsv(0, 0, 91%);
-@disabled-color         : fade(#000, 25%);
+@disabled-color         : @disabled;
 @slider-margin:                       14px 6px 10px;
-@component-background   : #fff;
-@slider-rail-background-color:        @background-color-base;
-@slider-rail-background-color-hover:  #e1e1e1;
-@slider-track-background-color:       @primary-3;
-@slider-track-background-color-hover: @primary-4;
-@slider-handle-color:                 @primary-3;
-@slider-handle-color-hover:           @primary-4;
-@slider-handle-color-focus:           tint(@primary-color, 20%);
-@slider-handle-color-focus-shadow:    tint(@primary-color, 50%);
+@component-background   : @white;
+@slider-rail-background-color:        @background;
+@slider-rail-background-color-hover:  @border;
+@slider-track-background-color:       @primary-color;
+@slider-track-background-color-hover: @primary-hover;
+@slider-handle-color:                 @primary-color;
+@slider-handle-color-hover:           @primary-hover;
+@slider-handle-color-focus:           @primary-focus;
+@slider-handle-color-focus-shadow:    @box-shadow-base;
 @slider-handle-color-tooltip-open:    @primary-color;
-@slider-dot-border-color:             @border-color-split;
-@slider-dot-border-color-active:      tint(@primary-color, 50%);
+@slider-dot-border-color:             @border;
+@slider-dot-border-color-active:      @primary-focus;
 @slider-disabled-color:               @disabled-color;
 @slider-disabled-background-color:    @component-background;
